Patch from NEST:
"I've made changes to the envisat reader to support versions 3H, 3K, 4A and 4B of ASAR products.
The ASA_WSS_IP has bands of different dimensions. I've changed bandinfo to handle this and it shouldn't have an affect on any other BEAM code.
In Product addBand() there is a throw if the band dimensions are not the same as the product dimensions. Is there other places that may make this assumption?"

 	Luis Veci



git-svn-id: https://www.brockmann-consult.de/svn/os/beam-4.x/trunk@1902 d90d2d48-2b33-0410-a973-e72f85796a02